requirements.txt
open_clip_torch
gradio<=3.44.0